otp/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-jalali-date==1.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jalali_core==1.0.0
jdatetime==5.0.0
PyJWT==2.10.1
sqlparse==0.5.3
